price

+++ shop_page/views.py
@@ -6,6 +6,11 @@ from project.settings import DATABASE
 from registration_page.models import User, Product
 from .app import shop
 from PIL import Image
+dict_types={
+    "IMG":"IMG",
+    "NAME":"TEXT",
+    "PRICE":"INT"
+}
 def render_shop_page():
     mod = False
     type1 = None
@@ -14,49 +19,71 @@ def render_shop_page():
         try:
             flask.request.form["send"].split(";")[1]
             type1 = flask.request.form["send"].split(";")[1]
+
             if type1 == "IMG":
                 id = int(flask.request.form["send"].split(";")[0])
                 path= os.path.abspath(__file__ + "/../static/image/"+ Product.query.all()[id-1].name+ ".png") 
-                os.remove(path)
+                
 
                 print(flask.request.form["send"])
 
                 # print(os.path.abspath(__file__+"/../static/image"))
                 print(path)
                 img = flask.request.files.get('image')
-                img.save(path)
+                print(img.filename, 11)
+                if img.filename != "":
+                    os.remove(path)
+                    img.save(path)
                 print(img, type(img))
                 # image = Image.open(img)
                 # print(image, type(image))
                 # image.show()
                 os.rename
-            elif type1== "TEXT" or type1 == "NAME":
-                text = flask.request.form["text"]
-                id = flask.request.form["send"].split(";")[0]
-                list_products = Product.query.all()
-                for product in Product.query.all():
-                    Product.query.filter_by(id = product.id).delete()
-                print(124)
-                for product_data in list_products:
-                    name = product_data.name
-                    print( product_data.id,id)
-                    if product_data.id == int(id):
-                        path1 = os.path.abspath(__file__ + "/../static/image/"+name+ ".png")
-                        path2 = os.path.abspath(__file__ + "/../static/image/"+text+ ".png")
-                        os.rename(path1,path2)
-                        name = text
-                        # print(123)
-                    # print(126)
-                    product  = Product(
-                        name= name ,
-                        description =product_data.description,
-                        count = product_data.count,
-                        price = product_data.price,
+            elif dict_types[type1] == "TEXT" or dict_types[type1] == "INT":
+                next = True
+                text = flask.request.form['text']
+                if dict_types[type1] == "INT":
+                    try:
+                        int(text)
+                    except:
+                        next = False
+                if text != "" and next:
+                    id = flask.request.form["send"].split(";")[0]
+                    list_products = Product.query.all()
+                    for product in Product.query.all():
+                        Product.query.filter_by(id = product.id).delete()
+                    print(124)
+                    for product_data in list_products:
+                        name = product_data.name
+                        count = product_data.count
+                        price = product_data.price
                         discount = product_data.discount
-                    )
-                    DATABASE.session.add(product)
-                print(125)
-                DATABASE.session.commit()
+                        print(name,price,discount, count)
+                        if product_data.id == int(id):
+                            if type1 == "NAME":
+
+                                path1 = os.path.abspath(__file__ + "/../static/image/"+name+ ".png")
+                                path2 = os.path.abspath(__file__ + "/../static/image/"+text+ ".png")
+                                os.rename(path1,path2)
+                                name = text
+                            elif type1 == "PRICE":
+                                price = flask.request.form['text']
+                            # print(123)
+                        # print(126)
+                        product  = Product(
+                            name= name ,
+                            description =product_data.description,
+                            count = count,
+                            price = price,
+                            discount = discount
+                        )
+                        DATABASE.session.add(product)
+                    print(125)
+                    DATABASE.session.commit()
+            # elif dict_types[type1] == "INT":
+            #     try:
+
+            #     except:
         except Exception as Error:
             print(Error)
             try:
@@ -69,7 +96,15 @@ def render_shop_page():
                     id = flask.request.form['name']
                     type1 = "NAME"
                 except:
-                    pass
+                    
+                    try:
+                        print(flask.request.form)
+                        id = flask.request.form['price']
+                        print('1')
+                        type1 = "PRICE"
+                        print('2')
+                    except:
+                        pass
             mod = True
     try:
         count =  len(flask.request.cookies.get('products').split(" "))
@@ -79,7 +114,8 @@ def render_shop_page():
         count = "0"
     # if flask.request.method == "POST":
         # count = str(int(count) + 1)
-    if len(list(Product.query.all())) == 0:
+    print(Product)
+    if len(Product.query.all()) == 0:
         path_excel=os.path.abspath(__file__ + "/../static/xlsx/Product.xlsx")
         read_excel = pandas.read_excel(io=path_excel,header=None,names=["name", "description","count","price","discount"])
         for row in read_excel.iterrows():
